divided code into smaller files
