+++ index.js
@@ -1,17 +1,17 @@
 require('dotenv').config();
 const express = require('express');
-const bodyParser = require('body-parser');
 const cors = require('cors');
+const bodyParser = require('body-parser');
+
 
 const { initializeDBConnection } = require('./db/db.connect');
 const { errorHandler } = require('./middleware/errorHandler');
 const { routeNotFound } = require('./middleware/routeNotFound');
 const { populateVideosCollection } = require('./utils/utils');
 
-const videos = require('./routes/video.router')
 
 const app = express();
-app.use(cors());
+app.use(cors({ origin: true }));
 app.use(express.urlencoded({
     extended: true
 }))
@@ -19,10 +19,10 @@ app.use(express.json());
 
 
 
+const videos = require('./routes/video.router')
+const playlist = require('./routes/playlist.router')
 
 initializeDBConnection();
-
-
 
 
 
@@ -30,7 +30,8 @@ app.get('/', (req, res) => {
     res.status(200).json({ "success": true, "message": "Hello Express App" })
 })
 
-app.use('/videos', videos)
+app.use('/videos', videos);
+app.use('/playlist', playlist)
 
 // populateVideosCollection();
 app.use(errorHandler);
